Lots of Changes

+++ portal/client/src/main/java/org/diskproject/client/application/hypothesis/HypothesisView.java
@@ -8,31 +8,38 @@ import org.diskproject.client.application.ApplicationSubviewImpl;
 import org.diskproject.client.components.hypothesis.HypothesisEditor;
 import org.diskproject.client.components.hypothesis.events.HypothesisSaveEvent;
 import org.diskproject.client.components.loader.Loader;
-import org.diskproject.client.components.treeold.TreeNode;
-import org.diskproject.client.components.treeold.TreeWidget;
-import org.diskproject.client.components.treeold.events.TreeItemDeletionEvent;
-import org.diskproject.client.components.treeold.events.TreeItemSelectionEvent;
+import org.diskproject.client.components.tloi.TriggeredLOIViewer;
+import org.diskproject.client.components.tree.TreeNode;
+import org.diskproject.client.components.tree.TreeWidget;
+import org.diskproject.client.components.tree.events.TreeItemActionEvent;
+import org.diskproject.client.components.tree.events.TreeItemSelectionEvent;
 import org.diskproject.client.place.NameTokens;
 import org.diskproject.client.rest.AppNotification;
 import org.diskproject.client.rest.DiskREST;
+import org.diskproject.shared.classes.common.Graph;
 import org.diskproject.shared.classes.common.TreeItem;
 import org.diskproject.shared.classes.hypothesis.Hypothesis;
+import org.diskproject.shared.classes.loi.TriggeredLOI;
 import org.diskproject.shared.classes.util.GUID;
 
 import com.google.gwt.core.client.Callback;
 import com.google.gwt.core.shared.GWT;
 import com.google.gwt.event.dom.client.ClickEvent;
+import com.google.gwt.event.dom.client.ClickHandler;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
 import com.google.gwt.uibinder.client.UiHandler;
 import com.google.gwt.user.client.History;
 import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.HTML;
+import com.google.gwt.user.client.ui.HTMLPanel;
 import com.google.gwt.user.client.ui.SimplePanel;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
 import com.vaadin.polymer.Polymer;
 import com.vaadin.polymer.elemental.Function;
+import com.vaadin.polymer.iron.widget.IronIcon;
+import com.vaadin.polymer.paper.widget.PaperButton;
 import com.vaadin.polymer.paper.widget.PaperFab;
 
 public class HypothesisView extends ApplicationSubviewImpl 
@@ -47,6 +54,7 @@ public class HypothesisView extends ApplicationSubviewImpl
   @UiField Loader loader;
   @UiField HypothesisEditor form;
   @UiField PaperFab addicon;
+  @UiField HTMLPanel matchlist;
   
   interface Binder extends UiBinder<Widget, HypothesisView> {
   }
@@ -54,6 +62,8 @@ public class HypothesisView extends ApplicationSubviewImpl
   @Inject
   public HypothesisView(Binder binder) {
     initWidget(binder.createAndBindUi(this));
+    tree.addCustomAction("query", null, "icons:explore", "green-button");
+    tree.addDeleteAction();
   }
   
   @Override
@@ -77,8 +87,11 @@ public class HypothesisView extends ApplicationSubviewImpl
     if(params.length == 0) {
       this.showHypothesisList();
     }
-    else {
+    else if(params.length == 1) {
       this.showHypothesis(params[0]);
+    }
+    else if(params.length == 2 && params[1].equals("query")) {
+      this.showHypothesisMatches(params[0]);
     }
   }
   
@@ -87,6 +100,7 @@ public class HypothesisView extends ApplicationSubviewImpl
     form.setVisible(false);
     tree.setVisible(false);
     addicon.setVisible(false);
+    matchlist.setVisible(false);
     addmode = false;
   }
   
@@ -106,20 +120,20 @@ public class HypothesisView extends ApplicationSubviewImpl
         AppNotification.notifyFailure(reason.getMessage());
         GWT.log("Failed", reason);
       }
-    });    
+    });
   }
   
   private void loadHypothesisTree(List<TreeItem> treelist) {
     TreeNode root = new TreeNode("Hypotheses", "Hypotheses", 
-        "A List of Hypotheses", false);
+        "A List of Hypotheses");
     
     HashMap<String, TreeNode> map = new HashMap<String, TreeNode>();
     for(TreeItem item : treelist) {
       TreeNode node = new TreeNode(
           item.getId(),
           item.getName(), 
-          item.getDescription(), true);
-      node.setDefaultIcon("icons:help-outline");
+          item.getDescription());
+      node.setIcon("icons:help-outline");
       node.setExpandedIcon("icons:help");
       node.setIconStyle("orange");
       map.put(item.getId(), node);
@@ -130,17 +144,13 @@ public class HypothesisView extends ApplicationSubviewImpl
       if(map.containsKey(item.getParentId()))
         parent = map.get(item.getParentId());
       
-      parent.addChild(node);
-      //tree.addNode(parent, node);
-      parent.setExpanded(true);
+      parent.setExpanded(true);      
+      tree.addNode(parent, node);
     }
-    tree.initialize(root);
-    //tree.setRoot(root);
+    tree.setRoot(root);
   }
   
   private void showHypothesis(final String id) {
-    tree.setVisible(false);
-    addicon.setVisible(false);
     loader.setVisible(true);
     Polymer.ready(form.getElement(), new Function<Object, Object>() {
       @Override
@@ -150,8 +160,8 @@ public class HypothesisView extends ApplicationSubviewImpl
           public void onSuccess(Hypothesis result) {
             loader.setVisible(false);
             form.setVisible(true);
-            form.load(result);
-            form.setNamespace(getNamespace(result.getId()));            
+            form.setNamespace(getNamespace(result.getId()));
+            form.load(result);            
           }
           @Override
           public void onFailure(Throwable reason) {
@@ -160,6 +170,74 @@ public class HypothesisView extends ApplicationSubviewImpl
           }
         });
         return null;
+      }
+    });
+  }
+  
+ private void showHypothesisMatches(final String id) {
+   loader.setVisible(true);   
+    DiskREST.queryHypothesis(id, 
+        new Callback<List<TriggeredLOI>, Throwable>() {
+      @Override
+      public void onSuccess(List<TriggeredLOI> result) {
+        loader.setVisible(false);
+        matchlist.setVisible(true);
+        showTriggeredLOIOptions(result);
+      }
+      @Override
+      public void onFailure(Throwable reason) {
+        loader.setVisible(false);
+        AppNotification.notifyFailure(reason.getMessage());
+      }
+    });
+  }
+
+  private void showTriggeredLOIOptions(List<TriggeredLOI> tlois) {
+    matchlist.clear();
+    
+    for(final TriggeredLOI tloi : tlois) {
+      final TriggeredLOIViewer tviewer = new TriggeredLOIViewer();
+      tviewer.initialize(userid, domain);
+      tviewer.load(tloi);
+      
+      /*ListNode node = new ListNode(tloi.getId(), tviewer);
+      node.setData(tloi);*/
+      /*node.setIcon("icons:explore");
+      node.setIconStyle("orange");
+      matchlist.addNode(node);*/
+      
+      final HTMLPanel panel = new HTMLPanel("");
+      panel.setStyleName("bordered-section");
+      panel.add(tviewer);
+      PaperButton button = new PaperButton("Trigger");
+      IronIcon icon = new IronIcon();
+      icon.setIcon("check");
+      button.add(icon);
+      
+      panel.add(button);
+      button.addClickHandler(new ClickHandler() {
+        @Override
+        public void onClick(ClickEvent event) {
+          triggerMatchedLOI(tloi);
+          matchlist.remove(panel);
+        }
+      });
+      matchlist.add(panel);
+    }
+  }
+  
+  void triggerMatchedLOI(final TriggeredLOI tloi) {
+    DiskREST.addTriggeredLOI(tloi, new Callback<Void, Throwable>() {
+      @Override
+      public void onFailure(Throwable reason) {
+        AppNotification.notifyFailure(reason.getMessage());
+      }
+      @Override
+      public void onSuccess(Void result) {
+        AppNotification.notifySuccess("Submitted.", 1000);
+        String token = NameTokens.getTLOIs()+"/" + userid+"/"
+            + domain + "/" + tloi.getId();
+        History.newItem(token, true);
       }
     });
   }
@@ -175,8 +253,9 @@ public class HypothesisView extends ApplicationSubviewImpl
 
     Hypothesis hypothesis = new Hypothesis();
     hypothesis.setId(id);
+    hypothesis.setGraph(new Graph());
+    form.setNamespace(this.getNamespace(id));    
     form.load(hypothesis);
-    form.setNamespace(this.getNamespace(id));
     
     History.newItem(this.getHistoryToken(id), false);
   }
@@ -212,26 +291,31 @@ public class HypothesisView extends ApplicationSubviewImpl
   @UiHandler("tree")
   void onTreeItemSelected(TreeItemSelectionEvent event) {
     TreeNode node = event.getItem();
-    String token = NameTokens.getHypotheses()+"/"+this.userid+"/"+this.domain;
-    token += "/" + node.getId();
+    String token = this.getHistoryToken(node.getId());
     History.newItem(token);
   }
   
   @UiHandler("tree")
-  void onTreeItemDeleted(TreeItemDeletionEvent event) {
+  void onTreeItemAction(TreeItemActionEvent event) {
     final TreeNode node = event.getItem();
-    if(Window.confirm("Are you sure you want to delete " + node.getName())) {
-      DiskREST.deleteHypothesis(node.getId(), new Callback<Void, Throwable>() {
-        @Override
-        public void onFailure(Throwable reason) {
-          AppNotification.notifyFailure(reason.getMessage());
-        }
-        @Override
-        public void onSuccess(Void result) {
-          AppNotification.notifySuccess("Deleted", 500);
-          tree.removeNode(node);
-        }
-      });
+    if(event.getAction().getId().equals("delete")) {
+      if(Window.confirm("Are you sure you want to delete " + node.getName())) {
+        DiskREST.deleteHypothesis(node.getId(), new Callback<Void, Throwable>() {
+          @Override
+          public void onFailure(Throwable reason) {
+            AppNotification.notifyFailure(reason.getMessage());
+          }
+          @Override
+          public void onSuccess(Void result) {
+            AppNotification.notifySuccess("Deleted", 500);
+            tree.removeNode(node);
+          }
+        });
+      }
+    }
+    else if(event.getAction().getId().equals("query")) {
+      String token = this.getHistoryToken(node.getId()) + "/query";
+      History.newItem(token);
     }
   }
